Refactor: SignIn and SignUp Refactoring and complete responsive design
