+++ src/Profile/SignUp/SignUp.module.css
@@ -1,3 +1,25 @@
-.SignUpFrame {
-    
+.SignUpMain {
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  padding: 20px;
+  align-items: center;
+}
+
+.signInButton {
+  background: rgba(169, 234, 195, 0.62);
+  border-radius: 0.625rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 8.75rem;
+  height: 2.5rem;
+  margin-top: 1rem;
+
+  color: rgba(0, 0, 0, 0.46);
+  text-align: left;
+  font-family: "Alef-Regular", sans-serif;
+  font-size: 1rem;
+  font-weight: 400;
+  cursor: pointer;
 }
