--- src/Profile/Signin/SignInButton.module.css
@@ -1,23 +1,0 @@
-.signInButton {
-  margin: -0.1 875rem 0 0 0;
-  background: var(--button-color, rgba(169, 234, 195, 0.62));
-  border-radius: 0.625rem;
-  display: flex;
-  flex-direction: row;
-  gap: 0.625rem;
-  align-items: center;
-  justify-content: center;
-  flex-shrink: 0;
-  width: 8.75rem;
-  height: 2.375rem;
-  position: relative;
-  cursor: pointer;
-  margin-top: 0.8rem;
-  margin-bottom: 10px;
-
-  color: rgba(0, 0, 0, 0.46);
-  text-align: left;
-  font-family: "Alef-Regular", sans-serif;
-  font-size: 1rem;
-  font-weight: 400;
-}
