+++ src/Profile/Signin/SignInInputField/SignInInputField.jsx
@@ -1,4 +1,4 @@
-import styles from "./Signin.module.css";
+import styles from "./SignInInputField.module.css";
 
 function SignInInputField({
   label,
@@ -8,8 +8,9 @@ function SignInInputField({
   errorMessage,
 }) {
   return (
-    <div className={styles.signInTextBox}>
-      <div className={styles.frame31}>
+    <div className={styles.inputFrame}>
+      <div className={styles.nameLabel}>{label}</div>
+      <div className={styles.input}>
         <input
           className={`${styles.placeholder} ${
             errorMessage ? styles.errorInput : ""
@@ -20,7 +21,6 @@ function SignInInputField({
           placeholder={errorMessage ? `⚠ ${errorMessage}` : label} // placeholder도 prop으로부터 받은 label을 사용합니다.
         />
       </div>
-      <div className={styles.nameLabel}>{label}</div>
     </div>
   );
 }
