+++ src/Profile/Signin/SignInInputField/SignInInputField.module.css
@@ -1,0 +1,74 @@
+.inputFrame {
+  flex-shrink: 0;
+  width: 20rem;
+  height: 4.5rem;
+  position: relative;
+  margin-bottom: 0.5rem;
+}
+
+.signInTextBox:focus {
+  border: none;
+}
+
+.nameLabel {
+  color: #282828;
+  text-align: left;
+  font-family: "Alef-Bold", sans-serif;
+  font-size: 1rem;
+  font-weight: 700;
+  position: relative;
+  margin-bottom: 0.3rem;
+}
+
+.input {
+  width: 100%;
+  height: 2.5rem;
+  position: relative;
+  background: #e9e9e9;
+  border-radius: 0.625rem;
+}
+
+.placeholder {
+  outline: rgba(54, 54, 54, 0.649);
+}
+
+.placeholder {
+  color: #979797;
+  text-align: left;
+  font-family: "Alef-Regular", sans-serif;
+  font-size: 0.9375rem;
+  font-weight: 400;
+  position: absolute;
+  padding-left: 0.875rem;
+  height: 100%;
+  width: 100%;
+}
+
+/* error */
+.errorInput {
+  border: 1.5px solid rgba(255, 0, 0, 0.745);
+  border-radius: 0.625rem;
+}
+
+
+@media (max-width: 1024px) {
+  .inputFrame {
+    width: 18rem; /* 288px */
+    height: 4rem; /* 64px */
+  }
+}
+
+@media (max-width: 768px) {
+  .inputFrame {
+    width: 16rem; /* 256px */
+    height: 3.5rem; /* 56px */
+    margin-bottom: 15px;
+  }
+}
+
+@media (max-width: 480px) {
+  .inputFrame {
+    width: 14rem; /* 224px */
+    height: 3rem; /* 48px */
+  }
+}
