+++ src/Profile/Signin/SignInMain.module.css
@@ -12,7 +12,7 @@
   font-family: "Alef-Regular", sans-serif;
   font-size: 2rem;
   font-weight: 400;
-  margin-bottom: 0.4rem;
+  margin-bottom: 1.5rem;
   position: relative;
   text-align: center;
 }
@@ -58,6 +58,7 @@
   display: flex;
   flex-direction: row;
   position: relative;
+  justify-content: center;
 }
 
 .AskSwitch {
@@ -74,4 +75,26 @@
   font-size: 1rem;
   font-weight: 400;
   cursor: pointer;
+}
+
+
+@media (max-width: 1024px) {
+  .signInFrame {
+    width: 25rem; /* 400px */
+    min-height: 31.25rem; /* 500px */
+  }
+}
+
+@media (max-width: 768px) {
+  .signInFrame {
+    width: 25rem; /* 320px */
+    min-height: 25rem; /* 400px */
+  }
+}
+
+@media (max-width: 480px) {
+  .signInFrame {
+    width: 20rem; /* 300px */
+    min-height: 21.875rem; /* 350px */
+  }
 }
