+++ src/Pureplate/Pureplate.module.css
@@ -1,8 +1,21 @@
-.homeLogedIn {
-  background: #ffffff;
-  height: 100%;
-  position: relative;
-  overflow: hidden;
+a,
+button,
+input,
+select,
+h1,
+h2,
+h3,
+h4,
+h5,
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+  border: none;
+  text-decoration: none;
+  appearance: none;
+  background: none;
+  -webkit-font-smoothing: antialiased;
 }
 
 .mapContainer {
